+++ YetiForce CRM 4.x.x/4.3.0_to_4.4.0/updates/files/app/TextParser.php
@@ -513,7 +513,7 @@ class TextParser
 		if (!isset($this->recordModel) || ($isPermitted && !Privilege::isPermitted($this->moduleName, 'DetailView', $this->record))) {
 			return '';
 		}
-		list($key, $params) = explode('|', $params, 2);
+		list($key, $params) = array_pad(explode('|', $params, 2), 2, false);
 		if ($this->recordModel->has($key)) {
 			$fieldModel = $this->recordModel->getModule()->getField($key);
 			if (!$fieldModel || !$this->useValue($fieldModel, $this->moduleName)) {
@@ -904,23 +904,27 @@ class TextParser
 	/**
 	 * Get last comments.
 	 *
-	 * @param int|bool $limit
+	 * @param mixed $params
 	 *
 	 * @return string
 	 */
-	protected function getComments($limit = false)
+	protected function getComments($params = false)
 	{
-		$query = (new \App\Db\Query())->select(['commentcontent'])->from('vtiger_modcomments')->where(['related_to' => $this->record])->orderBy(['modcommentsid' => SORT_DESC]);
+		list($limit, $showAuthor) = array_pad(explode('|', $params, 2), 2, false);
+		$query = (new \App\Db\Query())->select(['commentcontent', 'userid'])->from('vtiger_modcomments')->where(['related_to' => $this->record])->orderBy(['modcommentsid' => SORT_DESC]);
 		if ($limit) {
 			$query->limit($limit);
 		}
 		$commentsList = '';
-		foreach ($query->column() as $comment) {
-			if ($comment != '') {
-				$commentsList .= '<br /><br />' . nl2br($comment);
+		foreach ($query->all() as $comment) {
+			if ($comment['commentcontent'] != '') {
+				$commentsList .= '<br /><br />';
+				if ($showAuthor === 'true') {
+					$commentsList .= Purifier::encodeHtml(\App\Fields\Owner::getUserLabel($comment['userid'])) . ': ';
+				}
+				$commentsList .= nl2br($comment['commentcontent']);
 			}
 		}
-
 		return ltrim($commentsList, '<br /><br />');
 	}
 
@@ -949,7 +953,6 @@ class TextParser
 		if (isset($this->params[$params])) {
 			return $this->params[$params];
 		}
-
 		return '';
 	}
 
@@ -986,7 +989,6 @@ class TextParser
 		if ($instance->isActive()) {
 			return $instance->process();
 		}
-
 		return '';
 	}
 
@@ -1062,7 +1064,6 @@ class TextParser
 				}
 			}
 		}
-
 		return $variables;
 	}
 
@@ -1172,7 +1173,6 @@ class TextParser
 				$variables["$(custom : $fileName)$"] = Language::translate($instance->name, 'Other.TextParser');
 			}
 		}
-
 		return $variables;
 	}
 
@@ -1197,7 +1197,6 @@ class TextParser
 				}
 			}
 		}
-
 		return $variables;
 	}
 
@@ -1217,7 +1216,6 @@ class TextParser
 				'label' => Language::translate($relation->get('label'), $relation->get('relatedModuleName')),
 			];
 		}
-
 		return $variables;
 	}
 
@@ -1235,7 +1233,6 @@ class TextParser
 				'label' => Language::translate($module['name'], $module['name']),
 			];
 		}
-
 		return $variables;
 	}