+++ YetiForce CRM 4.x.x/4.3.0_to_4.4.0/updates/files/public_html/layouts/basic/modules/Vtiger/resources/AdvanceSearch.js
@@ -122,7 +122,7 @@ Vtiger_BasicSearch_Js("Vtiger_AdvanceSearch_Js", {
 		if (parent == false) {
 			return false;
 		}
-		var value = parent.find('.globalSearchValue').val();
+		var value = parent.find('.js-global-search__value').val();
 		if (value.length > 0) {
 			var form = this.getFilterForm();
 			var labelFieldList = this.getNameFields();
@@ -339,7 +339,7 @@ Vtiger_BasicSearch_Js("Vtiger_AdvanceSearch_Js", {
 			var currentElement = $(e.currentTarget);
 			currentElement.addClass('d-none');
 			var actionsContainer = currentElement.closest('.actions');
-			$('input[name="viewname"]', actionsContainer).removeClass('zeroOpacity').focus();
+			$('.js-name-filter', actionsContainer).removeClass('d-none').focus();
 			$('#advanceSave').removeClass('d-none');
 		});