+++ YetiForce CRM 4.x.x/4.3.0_to_4.4.0/updates/files/public_html/libraries/jQuery-Validation-Engine/js/jquery.validationEngine.js
@@ -774,8 +774,12 @@
 				options.showArrow = options.showArrowOnRadioAndCheckbox;
 			}
 
-			if(field.is(":hidden") && options.prettySelect && !field.is('textarea')) {
-				field = form.find("#" + options.usePrefix + methods._jqSelector(field.attr('id')) + options.useSuffix);
+			if (field.is(":hidden") && options.prettySelect && !field.is('textarea')) {
+				var jqSelector = methods._jqSelector(field.attr('id'));
+				field = form.find("#" + options.usePrefix + jqSelector);
+				if (field.length <= 0) {
+					field = form.find("#" + jqSelector + options.useSuffix);
+				}
 			}
 
 			if (options.isError && options.showPrompts){