+++ YetiForce CRM 4.x.x/4.3.0_to_4.4.0/updates/files/public_html/libraries/selectize/src/selectize.js
@@ -19,6 +19,7 @@ var Selectize = function($input, settings) {
 
 		eventNS          : '.selectize' + (++Selectize.count),
 		highlightedValue : null,
+		isBlurring       : false,
 		isOpen           : false,
 		isDisabled       : false,
 		isRequired       : $input.is('[required]'),
@@ -178,6 +179,7 @@ $.extend(Selectize.prototype, {
 		if ($input.attr('autocapitalize')) {
 			$control_input.attr('autocapitalize', $input.attr('autocapitalize'));
 		}
+		$control_input[0].type = $input[0].type;
 
 		self.$wrapper          = $wrapper;
 		self.$control          = $control;
@@ -185,6 +187,7 @@ $.extend(Selectize.prototype, {
 		self.$dropdown         = $dropdown;
 		self.$dropdown_content = $dropdown_content;
 
+		$dropdown.on('mouseenter mousedown click', '[data-disabled]>[data-selectable]', function(e) { e.stopImmediatePropagation(); });
 		$dropdown.on('mouseenter', '[data-selectable]', function() { return self.onOptionHover.apply(self, arguments); });
 		$dropdown.on('mousedown click', '[data-selectable]', function() { return self.onOptionSelect.apply(self, arguments); });
 		watchChildEvent($control, 'mousedown', '*:not(input)', function() { return self.onItemSelect.apply(self, arguments); });
@@ -360,7 +363,9 @@ $.extend(Selectize.prototype, {
 
 		// necessary for mobile webkit devices (manual focus triggering
 		// is ignored unless invoked within a click event)
-		if (!self.isFocused) {
+    // also necessary to reopen a dropdown that has been closed by
+    // closeAfterSelect
+		if (!self.isFocused || !self.isOpen) {
 			self.focus();
 			e.preventDefault();
 		}
@@ -648,10 +653,12 @@ $.extend(Selectize.prototype, {
 			// IE11 bug: element still marked as active
 			dest && dest.focus && dest.focus();
 
+			self.isBlurring = false;
 			self.ignoreFocus = false;
 			self.trigger('blur');
 		};
 
+		self.isBlurring = true;
 		self.ignoreFocus = true;
 		if (self.settings.create && self.settings.createOnBlur) {
 			self.createItem(null, false, deactivate);
@@ -989,7 +996,8 @@ $.extend(Selectize.prototype, {
 		return {
 			fields      : settings.searchField,
 			conjunction : settings.searchConjunction,
-			sort        : sort
+			sort        : sort,
+			nesting     : settings.nesting
 		};
 	},
 
@@ -1123,10 +1131,12 @@ $.extend(Selectize.prototype, {
 		$dropdown_content.html(html);
 
 		// highlight matching terms inline
-		if (self.settings.highlight && results.query.length && results.tokens.length) {
+		if (self.settings.highlight) {
 			$dropdown_content.removeHighlight();
-			for (i = 0, n = results.tokens.length; i < n; i++) {
-				highlight($dropdown_content, results.tokens[i].regex);
+			if (results.query.length && results.tokens.length) {
+				for (i = 0, n = results.tokens.length; i < n; i++) {
+					highlight($dropdown_content, results.tokens[i].regex);
+				}
 			}
 		}
 
@@ -1361,10 +1371,15 @@ $.extend(Selectize.prototype, {
 		self.loadedSearches = {};
 		self.userOptions = {};
 		self.renderCache = {};
-		self.options = self.sifter.items = {};
+		var options = self.options;
+		$.each(self.options, function(key, value) {
+			if(self.items.indexOf(key) == -1) {
+				delete options[key];
+			}
+		});
+		self.options = self.sifter.items = options;
 		self.lastQuery = null;
 		self.trigger('option_clear');
-		self.clear();
 	},
 
 	/**
@@ -1434,11 +1449,23 @@ $.extend(Selectize.prototype, {
 	 * @param {boolean} silent
 	 */
 	addItems: function(values, silent) {
+		this.buffer = document.createDocumentFragment();
+
+		var childNodes = this.$control[0].childNodes;
+		for (var i = 0; i < childNodes.length; i++) {
+			this.buffer.appendChild(childNodes[i]);
+		}
+
 		var items = $.isArray(values) ? values : [values];
 		for (var i = 0, n = items.length; i < n; i++) {
 			this.isPending = (i < n - 1);
 			this.addItem(items[i], silent);
 		}
+
+		var control = this.$control[0];
+		control.insertBefore(this.buffer, control.firstChild);
+
+		this.buffer = null;
 	},
 
 	/**
@@ -1491,13 +1518,16 @@ $.extend(Selectize.prototype, {
 				// hide the menu if the maximum number of items have been selected or no options are left
 				if (!$options.length || self.isFull()) {
 					self.close();
-				} else {
+				} else if (!self.isPending) {
 					self.positionDropdown();
 				}
 
 				self.updatePlaceholder();
 				self.trigger('item_add', value, $item);
-				self.updateOriginalInput({silent: silent});
+
+				if (!self.isPending) {
+					self.updateOriginalInput({silent: silent});
+				}
 			}
 		});
 	},
@@ -1752,7 +1782,13 @@ $.extend(Selectize.prototype, {
 
 		if (self.settings.mode === 'single' && self.items.length) {
 			self.hideInput();
-			self.$control_input.blur(); // close keyboard on iOS
+
+			// Do not trigger blur while inside a blur event,
+			// this fixes some weird tabbing behavior in FF and IE.
+			// See #1164
+			if (!self.isBlurring) {
+				self.$control_input.blur(); // close keyboard on iOS
+			}
 		}
 
 		self.isOpen = false;
@@ -1773,7 +1809,7 @@ $.extend(Selectize.prototype, {
 		offset.top += $control.outerHeight(true);
 
 		this.$dropdown.css({
-			width : $control.outerWidth(),
+			width : $control[0].getBoundingClientRect().width,
 			top   : offset.top,
 			left  : offset.left
 		});
@@ -1809,11 +1845,15 @@ $.extend(Selectize.prototype, {
 	 */
 	insertAtCaret: function($el) {
 		var caret = Math.min(this.caretPos, this.items.length);
+		var el = $el[0];
+		var target = this.buffer || this.$control[0];
+
 		if (caret === 0) {
-			this.$control.prepend($el);
+			target.insertBefore(el, target.firstChild);
 		} else {
-			$(this.$control[0].childNodes[caret]).before($el);
+			target.insertBefore(el, target.childNodes[caret]);
 		}
+
 		this.setCaret(caret + 1);
 	},
 
@@ -2049,6 +2089,11 @@ $.extend(Selectize.prototype, {
 		self.$control_input.removeData('grow');
 		self.$input.removeData('selectize');
 
+		if (--Selectize.count == 0 && Selectize.$testInput) {
+			Selectize.$testInput.remove();
+			Selectize.$testInput = undefined;
+		}
+
 		$(window).off(eventNS);
 		$(document).off(eventNS);
 		$(document.body).off(eventNS);
@@ -2091,11 +2136,16 @@ $.extend(Selectize.prototype, {
 
 		// add mandatory attributes
 		if (templateName === 'option' || templateName === 'option_create') {
-			html.attr('data-selectable', '');
+			if (!data[self.settings.disabledField]) {
+				html.attr('data-selectable', '');
+			}
 		}
 		else if (templateName === 'optgroup') {
 			id = data[self.settings.optgroupValueField] || '';
 			html.attr('data-group', id);
+			if(data[self.settings.disabledField]) {
+				html.attr('data-disabled', '');
+			}
 		}
 		if (templateName === 'option' || templateName === 'item') {
 			html.attr('data-value', value || '');